chore: moved the builder to the appropriate path
Moved builder to appropriate path and changed tasks. Updated swagger definition
